podcast_generation/requirements.txt
openai
fastapi
mangum
uvicorn
python-multipart
supabase
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
httpx
requests
sqlmodel